Use in-publish to workaround broken npm
npm has some really broken behaviour - on npm install
it runs the prepublish script. This often depends on
packages that should be installed during npm install, leading to
an ugly error for the user.

in-install is a package that works around this

see https://github.com/npm/npm/issues/3059 for discussion of this problem